+++ src/components/registerForm/cardNumber/CardNumber.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta } from "@storybook/react";
-import React, { useRef } from "react";
-import { CardNumber } from ".";
+import React from "react";
+import CardNumber from ".";
 
 const cardNumber = {
   component: CardNumber,
@@ -10,20 +10,13 @@ const cardNumber = {
 export default cardNumber;
 
 export const Example = () => {
-  const divRef = useRef<HTMLDivElement>(null);
-
   const registerCard = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
-    divRef.current?.childNodes.forEach((elem) => {
-      if (elem instanceof HTMLInputElement) {
-        console.log(elem.value);
-      }
-    });
   };
 
   return (
     <form onSubmit={registerCard}>
-      <CardNumber ref={divRef} />
+      <CardNumber />
       <button>다음</button>
     </form>
   );
