+++ src/components/registerForm/expireDate/index.tsx
@@ -4,10 +4,10 @@ import styled, { css } from "styled-components";
 import ErrorSpan from "src/components/@common/ErrorSpan";
 import FormLabel from "src/components/@common/FormLabel";
 import { MMYY_REGEXP, ONLY_NUMBER_REGEXP } from "src/utils/regexp";
-import { InputValuesContext } from "../InputValueContext";
+import { inputValuesContext } from "src/InputValueContext";
 
 function ExpireDate() {
-  const [cardInput, setCardInput] = useContext(InputValuesContext);
+  const [cardInput, setCardInput] = useContext(inputValuesContext);
 
   const [expireError, setExpireError] = useState(false);
