+++ src/components/registerForm/securityCode/index.tsx
@@ -4,10 +4,10 @@ import FormLabel from "src/components/@common/FormLabel";
 import Input from "src/components/@common/Input";
 import { ONLY_NUMBER_REGEXP } from "src/utils/regexp";
 import styled, { css } from "styled-components";
-import { InputValuesContext } from "../InputValueContext";
+import { inputValuesContext } from "src/InputValueContext";
 
 function SecurityCode() {
-  const [cardInput, setCardInput] = useContext(InputValuesContext);
+  const [cardInput, setCardInput] = useContext(inputValuesContext);
   const [error, setError] = useState(false);
 
   const codeChange: React.ChangeEventHandler<HTMLInputElement> = (event) => {
