+++ src/interfaces/index.ts
@@ -1,0 +1,6 @@
+export interface CardNumberObj {
+  first: string;
+  second: string;
+  third: string;
+  fourth: string;
+}
